Refactor Settings
Refactor app to sync settings to a thread-safe struct so that
the log file debounce handler can apply new settings immediately.
Does not apply to settings used for initialization, such as the
API key and usernames.

Refactor so that settings are automatically savedonce three
seconds have passed since the last change.

Moved Start/Stop button to be between initialization and runtime
settings. It would be better to make the sections more distinct
so that the fact that the initialization settings require a
restart is more obvious.